requirements.txt
python-rocksdb
pickle